requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
Django==4.2.11
django-allauth==0.61.1
django-filter==24.2
idna==3.7
oauthlib==3.2.2
pip==23.2.1
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==68.2.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.41.2
